requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
Django==1.8.2
django-filter==0.12.0
djangorestframework==3.3.2
kombu==3.0.26
Markdown==2.6.5
psycopg2==2.6.1
pytz==2015.4
redis==2.10.3
requests==2.9.1
